BREAKING CHANGE(web-react): Remove feature flag for the uniform `Modal` variant #DS-1181
The uniform appearance of `Modal` is now default. The docked variant can be
turned on using the `isDockedOnMobile` option.
